+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v3.0.1
 
 - Fix Android 15 Easter Egg add flag crash on Android 8.1 [#384](https://github.com/hushenghao/AndroidEasterEggs/issues/384)
+- Fix Android M Easter Egg players color [#387](https://github.com/hushenghao/AndroidEasterEggs/issues/387)
 - Upgrade project dependencies
 
 ### v3.0.0 (2024-09-18)